remove liq
