+++ dex/pair/wasm/src/lib.rs
@@ -21,8 +21,6 @@ multiversx_sc_wasm_adapter::endpoints! {
     (
         init => init
         upgrade => upgrade
-        removeLiquidity => remove_liquidity
-        removeLiquidityAndBuyBackAndBurnToken => remove_liquidity_and_burn_token
         swapNoFeeAndForward => swap_no_fee
         swapTokensFixedInput => swap_tokens_fixed_input
         swapTokensFixedOutput => swap_tokens_fixed_output
@@ -75,6 +73,8 @@ multiversx_sc_wasm_adapter::endpoints! {
         getState => state
         addInitialLiquidity => add_initial_liquidity
         addLiquidity => add_liquidity
+        removeLiquidity => remove_liquidity
+        removeLiquidityAndBuyBackAndBurnToken => remove_liquidity_and_burn_token
     )
 }
